--- packages/@react-spectrum/card/README.md
@@ -1,3 +1,0 @@
-# @react-spectrum/card
-
-This package is part of [react-spectrum](https://github.com/adobe/react-spectrum). See the repo for more details.
