--- packages/@react-spectrum/card/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@react-spectrum/card",
-  "version": "3.0.0-alpha.1",
-  "private": true,
-  "description": "Spectrum UI components in React",
-  "license": "Apache-2.0",
-  "main": "dist/main.js",
-  "module": "dist/module.js",
-  "types": "dist/types.d.ts",
-  "source": "src/index.ts",
-  "files": ["dist", "src"],
-  "sideEffects": [
-    "*.css"
-  ],
-  "targets": {
-    "main": {
-      "includeNodeModules": ["@adobe/spectrum-css-temp"]
-    },
-    "module": {
-      "includeNodeModules": ["@adobe/spectrum-css-temp"]
-    }
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/adobe/react-spectrum"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.6.2",
-    "@react-aria/card": "3.0.0-alpha.1",
-    "@react-aria/utils": "^3.0.0",
-    "@react-spectrum/utils": "^3.0.0",
-    "@react-stately/card": "3.0.0-alpha.1",
-    "@react-types/card": "3.0.0-alpha.1",
-    "@react-types/shared": "^3.0.0"
-  },
-  "devDependencies": {
-    "@adobe/spectrum-css-temp": "3.0.0-alpha.1"
-  },
-  "peerDependencies": {
-    "react": "^16.8.0 || ^17.0.0-rc.1",
-    "@react-spectrum/provider": "^3.0.0-rc.1"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
