+++ packages/@react-spectrum/cards/intl/ar-AE.json
