+++ packages/@react-spectrum/cards/intl/bg-BG.json
