+++ packages/@react-spectrum/cards/intl/de-DE.json
