+++ packages/@react-spectrum/cards/intl/el-GR.json
