+++ packages/@react-spectrum/cards/intl/en-US.json
