+++ packages/@react-spectrum/cards/intl/fr-FR.json
