+++ packages/@react-spectrum/cards/intl/he-IL.json
