+++ packages/@react-spectrum/cards/intl/hu-HU.json
