+++ packages/@react-spectrum/cards/intl/ja-JP.json
