+++ packages/@react-spectrum/cards/intl/ko-KR.json
