+++ packages/@react-spectrum/cards/intl/nb-NO.json
