+++ packages/@react-spectrum/cards/intl/sk-SK.json
