+++ packages/@react-spectrum/cards/intl/sl-SI.json
