+++ packages/@react-spectrum/cards/intl/tr-TR.json
