+++ packages/@react-spectrum/cards/intl/uk-UA.json
