+++ packages/@react-spectrum/cards/intl/zh-CN.json
