+++ packages/@react-spectrum/cards/src/index.ts
@@ -12,4 +12,7 @@
 
 /// <reference types="css-module-types" />
 
+export * from './CardView';
+export * from './BaseLayout';
+export * from './GridLayout';
 export * from './Card';
